Firewall: Rules/NAT: Invert Port Match
This code is for demonstration only.  To show how an invert port match feature
could be implemented in OPNsense.

This code has not been evaluated nor tested to be safe for implementation or
use.  This code may contain flaws that result in unexpected behavior and
failure.  Implementation and use of this code, whether in part or in whole,
is at the sole risk of the implementer/user.

Dependent: "Interfaces: Block Traffic Targeting NOT Enabled TCP/UDP Ports - Filter Option"