Interfaces: Block Traffic Targeting NOT Enabled TCP/UDP Ports - Filter Option
This code is for demonstration only.  To show how an interface option to
immediately block inbound traffic targeting TCP/UDP ports that are not enabled*
could be implemented in OPNsense.

This code has not been evaluated nor tested to be safe for implementation or
use.  This code may contain flaws that result in unexpected behavior and
failure.  Implementation and use of this code, whether in part or in whole,
is at the sole risk of the implementer/user.

Immediately blocking this traffic that has no possibility of matching a
firewall rule reduces processing unnecessary further evaluations and logging.
* Destination ports of enabled inbound firewall rules.

Dependent on "Firewall: Rules/NAT: Invert Port Match"